Implemented decay for current highscores over time to ensure they gradually decrease as their associated thresholds become outdated. This allows new, lower scores to surpass them as the top score, ensuring strategies remain aligned with the current market conditions, even when prices (and potential scores) decline.

+++ BitBetMatic/Backtesting/PortfolioManager.cs
@@ -1,11 +1,10 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using BitBetMatic;
 public class PortfolioManager
 {
     private decimal _cashBalance;
-    private decimal tradeMargin = (decimal)0.9975;
+    private decimal tradeMargin = 0.9975m;
     private Dictionary<string, (decimal tokenAmount, decimal currentPrice)> _assetBalances;
 
     public PortfolioManager()
